+++ src/index.ts
@@ -1,5 +1,7 @@
+export { BackendResult } from './backend-result';
+export { Query } from './queries/query';
 export { QueryBuilder } from './query-builder';
 export { QueryResult } from './query-result';
-export { BackendResult } from './backend-result';
+export { QueryRunner } from './query-runner';
 export * as queries from './queries/index';
 export * as interfaces from './support/index';