+++ changelog.md
@@ -2,6 +2,14 @@
 
 All notable changes to `LaravelQueryApiFrontend` will be documented in this file.
 
+## Version 0.0.3
+
+### Added
+
+- QueryRunner class
+
+- Query class to index file
+
 ## Version 0.0.2
 
 ### Added